+++ src/Database/QueryBuilder/QueryBuilderInterface.php
@@ -10,35 +10,35 @@ interface QueryBuilderInterface
      * @param MigrationTable $table
      * @return array list of queries
      */
-    public function createTable(MigrationTable $table);
+    public function createTable(MigrationTable $table): array;
 
     /**
      * @param MigrationTable $table
      * @return array list of queries
      */
-    public function dropTable(MigrationTable $table);
+    public function dropTable(MigrationTable $table): array;
 
     /**
      * @param MigrationTable $table
      * @return array list of queries
      */
-    public function renameTable(MigrationTable $table);
+    public function renameTable(MigrationTable $table): array;
 
     /**
      * @param MigrationTable $table
      * @return array list of queries
      */
-    public function alterTable(MigrationTable $table);
+    public function alterTable(MigrationTable $table): array;
 
     /**
      * @param MigrationTable $table
      * @return array list of queries
      */
-    public function copyTable(MigrationTable $table);
+    public function copyTable(MigrationTable $table): array;
 
     /**
      * @param string $string
      * @return string escaped string
      */
-    public function escapeString($string);
+    public function escapeString(string $string): string;
 }