+++ src/Migration/FilesFinder.php
@@ -8,47 +8,47 @@ use Nette\Utils\Finder;
 class FilesFinder
 {
     private $directories = [];
-    
-    public function addDirectory($path)
+
+    public function addDirectory(string $path): FilesFinder
     {
         if (!file_exists($path)) {
             throw new InvalidArgumentException('Directory "' . $path . '" not found');
         }
-        
+
         if (!is_dir($path)) {
             throw new InvalidArgumentException('"' . $path . '" is not directory');
         }
-        
+
         $index = array_search($path, $this->directories);
         if ($index === false) {
             $this->directories[] = $path;
         }
         return $this;
     }
-    
-    public function removeDirectory($path)
+
+    public function removeDirectory(string $path): FilesFinder
     {
         $index = array_search($path, $this->directories);
         if ($index === false) {
             throw new InvalidArgumentException('"Path ' . $path . '" was not found in list of directories');
         }
-        
+
         unset($this->directories[$index]);
         return $this;
     }
-    
-    public function removeDirectories()
+
+    public function removeDirectories(): FilesFinder
     {
         $this->directories = [];
         return $this;
     }
-    
-    public function getDirectories()
+
+    public function getDirectories(): array
     {
         return $this->directories;
     }
-    
-    public function getFiles()
+
+    public function getFiles(): array
     {
         $files = [];
         foreach ($this->directories as $directory) {