+++ testing_migrations/phoenix/20170705144807_copy_table.php
@@ -7,14 +7,14 @@ use Phoenix\Migration\AbstractMigration;
 
 class CopyTables extends AbstractMigration
 {
-    public function up()
+    public function up(): void
     {
         $this->table('table_2')->copy('new_table_2');
         $this->table('table_3')->copy('new_table_3', MigrationTable::COPY_STRUCTURE_AND_DATA);
         $this->table('table_2')->copy('new_table_2', MigrationTable::COPY_ONLY_DATA);
     }
 
-    protected function down()
+    protected function down(): void
     {
         $this->table('new_table_3')->drop();
         $this->table('new_table_2')->drop();